refactor: fused moe to nomic
